requirements.txt
azure-identity==1.16.0
azure-ai-ml==1.15.0
scikit-learn==1.5.1
joblib==1.4.2
numpy==1.26.4
fastapi==0.115.0
uvicorn==0.30.6
mlflow==2.14.1
pytest==8.0.0
pytest-cov==5.0.0
google-cloud-aiplatform>=1.50.0
google-cloud-logging>=3.11.0
